refactor(S1): actions 폴더 정리
